add isBetExisting check to all queries requiring an ID

+++ backend/queries/bets/bets.js
@@ -29,12 +29,17 @@ const getBets = async (req, res, next) => {
 const getBetsById = async (req,res,next) => {
     let {betId} = req.params
     try{ 
-        let bet = await db.any("SELECT from bets WHERE id=$1 INNER JOIN users ON bets.better_id=users.id", betId);
-        res.status(200).json({
-            bet,
-            status: "sucess",
-            message: "Grabbed bet by ID"
-        })
+        if(await isBetExisting(betId)) {
+            let bet = await db.any("SELECT from bets WHERE id=$1 INNER JOIN users ON bets.better_id=users.id", betId);
+            res.status(200).json({
+                bet,
+                status: "sucess",
+                message: "Grabbed bet by ID"
+            })
+        } else {
+
+        }
+        
     } catch(err) {
         next(err)
     }
@@ -69,12 +74,18 @@ const postBets = async (req,res,next) => {
 const patchBets = async (req,res,next) => {
     try{
         let {taker_id} = req.body;
-        let updateBet = await db.one("UPDATE bets SET taker_id=$1 RETURNING *", taker_id);
-        res.status(200).json({
-            updateBet,
-            status:"Sucess",
-            message: "Updated bet"
-        })
+        let {betId} = req.params;
+        if(await isBetExisting(betId)) {
+            let updateBet = await db.one("UPDATE bets SET taker_id=$1 WHERE id=$2 RETURNING *", [taker_id, betId]);
+            res.status(200).json({
+                updateBet,
+                status:"Sucess",
+                message: "Updated bet"
+            })
+        } else {
+
+        }
+        
     } catch(err) {
         next(err)
     }
@@ -83,12 +94,15 @@ const patchBets = async (req,res,next) => {
 const deleteBet = async (req,res,next) => {
     try{
         let {betId} = req.params;
-        let eraseBet = await db.one("DELETE FROM bets WHERE id=$1 RETURNING *", betId);
-        res.status(200).json({
-            eraseBet,
-            status: "Sucess",
-            message: "Deleted Bet"
-        })
+        if(await isBetExisting(betId)) {
+            let eraseBet = await db.one("DELETE FROM bets WHERE id=$1 RETURNING *", betId);
+            res.status(200).json({
+                eraseBet,
+                status: "Sucess",
+                message: "Deleted Bet"
+            })
+        }
+        
 
     } catch(err){
         next(err)
